+++ services/supplier.service.js
@@ -1,4 +1,5 @@
 import supplierRepository from "../repositories/supplier.repository.js";
+import productRepository from "../repositories/product.repository.js";
 
 async function createSupplier(supplier){
     return await supplierRepository.insertSupplier(supplier);
@@ -17,6 +18,13 @@ async function updateSupplier(supplier){
 }
 
 async function deleteSupplier(id){
+
+    const products = await productRepository.getProductsBySupplierId(id);
+    
+    
+    if(products) {
+        throw new Error("Supplier can't be deleted. There are products from this supplier");
+    }
     return await supplierRepository.deleteSupplier(id);
 }
